increase ticker resolution:
- Before we used only millisecond resolution for ticker interval, which
  would break for high values of the speed flag.  Here we change to use
  nanosecond resolution enabling much faster playback.

+++ main_test.go
@@ -10,7 +10,7 @@ import (
 
 type outputSpeedTest struct {
 	input            []string
-	speed            int
+	speed            int64
 	expectedDuration time.Duration
 }
